% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{boreal_caribou}
\alias{boreal_caribou}
\title{Boreal caribou (\emph{Rangifer tarandus}) movement data}
\format{
Data frame with columns for ID, Time (POSIX UTC),
X,Y spatial coordinates (ESPG code 32611), and season.
}
\source{
GNWT
}
\usage{
data(boreal_caribou)
}
\description{
N= 10 anonymized caribou, 2015-2017, with an 8 hr median sampling rate and
annotated by relevant ecological seasons.
}
\examples{
library(sf);library(mapview);library(dplyr)

# load caribou data
data(boreal_caribou)

# load highway data
data(hwy1_west)

# convert to sf structure
boreal_caribou_sf <- st_as_sf(boreal_caribou, 
                               coords = c("X","Y"), crs = 32611)

# map track highway data
boreal_tracks <- boreal_caribou_sf |> 
  group_by(ID) |>
  summarize(do_union=FALSE) |> 
  st_cast("LINESTRING")

mapview(boreal_tracks, zcol = "ID") + 
  mapview(hwy1_west, color = "red")

}
\keyword{data}
