% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePermFits.R
\name{comparePermFits}
\alias{comparePermFits}
\title{compare Permeability Model Fits}
\usage{
comparePermFits(formulae, models = NULL, data)
}
\arguments{
\item{formulae}{list of formulae, with the structure "~ covariate..."}

\item{models}{optional - a list of fitted models, as from
\link{fitPermeability}}

\item{data}{crossing table, as output from \link{buildCrossingTable}}
}
\value{
A dataframe with rows for each model, columns for AIC,
the delta AIC ("dAIC" or difference in AIC between consecutive models),
number of parameters (\code{k}), log Likelihood value, and
sorted by AIC values.
}
\description{
Given a list of models or model formulae, compares model
fits using AIC values. Output is ranked by the lowest AIC to highest
AIC model.
}
\examples{
require(permeability)
require(lubridate)

data("example_cT")

example_cT$doy <- lubridate::yday(example_cT$Date)

model_formulae <- list(null = ~ 1, 
                       Traffic = ~ traffic_volume, 
                       Season = ~ Season, 
                       sDOY = ~ s(doy, k = 5, bs = "cc"),
                       Traffic_Season = ~ traffic_volume + Season)

comparePermFits(model_formulae, data = example_cT)
}
