% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{example_cT}
\alias{example_cT}
\title{Example crossing table}
\format{
An object of class \code{data.frame} with 191 rows and 14 columns.
}
\usage{
data(example_cT)
}
\description{
Example crossing table, as the result of running the examples in the
BuildCrossingTable function.
}
\examples{
library(sf);library(mapview);library(dplyr)

# load caribou data
data(boreal_caribou)

# load highway data
data(hwy1_west)

# convert to sf structure
boreal_caribou_sf <- st_as_sf(boreal_caribou, 
                               coords = c("X","Y"), crs = 32611)

# map track highway data
boreal_tracks <- boreal_caribou_sf |> 
  group_by(ID) |>
  summarize(do_union=FALSE) |> 
  st_cast("LINESTRING")

mapview(boreal_tracks, zcol = "ID") + 
  mapview(hwy1_west, color = "red")

}
\keyword{data}
