% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPermeability.R
\name{fitPermeability}
\alias{fitPermeability}
\title{Estimate permeability}
\usage{
fitPermeability(formula = ~1, data, beta0 = NULL)
}
\arguments{
\item{formula}{one-sided formula - uses names of columns in the crossing
table as possible covariates. Defaults to NULL (null model).}

\item{data}{crossing table object - output of
\code{\link{build_crossingTable}}.}

\item{beta0}{optional parameter for initial values for model to be optimized
over. Default NULL (see \code{optim} function for more details).}
}
\description{
Estimates permeability using the crossing table as output from
\link{buildCrossingTable}. If no formula is provided, will default to a
null model of permeability, returning one value for kappa. Otherwise, if
covariate values are provided in the \code{formula}, will fit a model that
estimates permeability as a logit-linear function of covariate(s), returning
a kappa value for each unique covariate value. Models are fit using
maximum likelihood estimation, with the \code{optim} function.
}
\examples{
require(permeability)
data("example_cT")

# NULL MODEL ###########################################
null_model <- fitPermeability(data = example_cT)
print(null_model)

# COVARIATE MODEL ##################################

## fixed effect
traffic_model <- fitPermeability(~ traffic_volume, data = example_cT)

## get predictions with 95\% CI using bootstrapping
predict_kappa <- predict(traffic_model, se.fit = TRUE)
predict_kappa$traffic_volume <- traffic_model$data$traffic_volume
predict_kappa <- predict_kappa |>
  dplyr::arrange(kappa.hat)

plot(predict_kappa$traffic_volume, predict_kappa$kappa.hat, type = "l", 
     xlab = "Daily Traffic Volume", 
     ylab = "Estimated Kappa Values", 
     col ="orange", lwd = 2)

## day of year smoother
example_cT$doy <- lubridate::yday(example_cT$Date)
doy_model <- fitPermeability(~ s(doy, k = 6, bs = "cc"), data = example_cT)
doy_model

predict_kappa <- predict(doy_model, se.fit = TRUE, bootstrap = TRUE,
                         new.data = data.frame(doy = seq(min(example_cT$doy),
                                                         max(example_cT$doy),
                                                         by = 1)))
predict_kappa$doy <- seq(min(example_cT$doy), max(example_cT$doy),
                         by = 1)
}
