% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{hwy1_west}
\alias{hwy1_west}
\title{Highway 1 West}
\format{
Linear simple feature
}
\source{
Extracted and processed from the Road Network File - 2005 -
Northwest Territories
(https://open.canada.ca/data/en/dataset/97adfa76-7eee-477b-b3d3-e430be7bd812)
}
\usage{
data(hwy1_west)
}
\description{
Highway 1 West
}
\examples{
library(sf);library(mapview);library(dplyr)

# load caribou data
data(boreal_caribou)

# load highway data
data(hwy1_west)

# convert to sf structure
boreal_caribou_sf <- st_as_sf(boreal_caribou, 
                               coords = c("X","Y"), crs = 32611)

# map track highway data
boreal_tracks <- boreal_caribou_sf |> 
  group_by(ID) |>
  summarize(do_union=FALSE) |> 
  st_cast("LINESTRING")

mapview(boreal_tracks, zcol = "ID") + 
  mapview(hwy1_west, color = "red")

}
\keyword{data}
