% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPredictions.R
\name{mapPredictions}
\alias{mapPredictions}
\title{Map Predictions}
\usage{
mapPredictions(
  predictions,
  barrier.sf,
  interactive = FALSE,
  tracks.sf = NULL,
  breaks = NULL,
  add.basemap = FALSE,
  basemap = NULL,
  box = NULL,
  map.service = "esri",
  map.type = "world_imagery",
  map.res = NULL,
  nocrossings.color = "grey"
)
}
\arguments{
\item{predictions}{dataframe, at minimum, with predicted kappa values and
a \code{barrier.id} column to identify distinct barrier segments. This column
corresponds to the same \code{barrier.id} values in the \code{permdata} \code{barrier}
object and crossing table object (as output from \code{buildCrossingTable}). If
there are multiple linear features, these should be identified with an
additional \code{Feature.ID} column; this column will be used to correspond with
a matching column in the provided \code{barrier.sf} object.}

\item{barrier.sf}{sf LINESTRING object(s) of the barrier. Should be the same
barrier object(s) used as input to the \code{prepPermeability} function. Can
optionally have a \code{Feature.ID} column, which should have unique values
that can be used to match with a similar column in the \code{predictions} object.}

\item{interactive}{if TRUE, will return an interactive map using the
\code{mapview} package. Default FALSE.}

\item{tracks.sf}{optionally, an sf POINT object of animal track locations,
with a columns for \code{ID}. Will be used to map animal tracks behind the
barrier and should have a matching CRS to the \code{barrier.sf} object.}

\item{breaks}{Default NULL but if specified, should be a vector of values
to break up the predicted kappa values in the plot legend. If NULL, will
use a log scale to break up data and plot it. Applicable for static maps
only.}

\item{add.basemap}{if TRUE (default FALSE), will add a basemap for a
static map, using an ESRI landscape map from the \code{basemaps} package.
Alternatively you can provide your own basemap (as a raster) with the
\code{basemap} argument.}

\item{basemap}{default NULL but you can prove your own basemap for plotting
here. Should be a raster with similar spatial extent and CRS as your
\code{barrier.sf} object. Applicable for static maps only.}

\item{box}{default NULL but you can provide a different spatial extent for
the map using this box parameter, which should be an sf bbox object with a
CRS matching the barrier sf object. If NULL, will use the spatial extent of
the provided barrier object to create map. Applicable to static maps only.}

\item{map.service}{for static maps with an automatic basemap (provided
through the "basemaps" package), the map service to use - default is "esri"
but you can also use other options available through the "basemap_raster"
function from basemaps (e.g., "osm").}

\item{map.type}{for static maps with an automatic basemap (provided
through the "basemaps" package), the map type to use - default is
"world_imagery" from ESRI but you can provide any other options
available through the "get_maptypes" function from the basemaps package
(just provide your map service type).}

\item{map.res}{for static maps with an automatic basemap (provided
through the "basemaps" package), the resolution of the map (0-1). Default
NULL (will use the basemaps default).}

\item{nocrossings.color}{for static maps and portions of the road with no
actual or null
crossings, what color to show these segments on the road sf object.}
}
\value{
An interactive or static map of predicted kappa values for barrier
segments.
}
\description{
Function to map kappa predictions as output from the \code{predict}
function and a fitted permeability model from \code{fitPermeability}, to
segments of a barrier \code{sf} object. By default uses the \code{ggplot2} and
\code{ggspatial} packages to create a static map (with optional basemap).
Users can also create an interactive map, using the \code{mapview} package.
}
\examples{
require(permeability)
require(ggplot2)
require(ggspatial)
require(mapview)
require(basemaps)

data("example_id")
data("example_cT")
data("hwy1_west")

# fit model
hwy1_w_model <- fitPermeability(~ s(km, k = 4, bs = "cr"),
                                     data = example_cT)

# get predictions with barrier.id as column
hwy1_w_xy <- st_coordinates(hwy1_west)
colnames(hwy1_w_xy) <- c("X","Y","line_id")
hwy1_w_z <- hwy1_w_xy[,1] + 1i*hwy1_w_xy[,2]

predictions <- cbind(hwy1_w_xy, hwy1_w_z) |> 
  data.frame() |>
  group_by(line_id) |>
  reframe(km = Mod(diff(hwy1_w_z))) |>
  select(km) |>
  mutate(km = cumsum(km)/1000) |>
  mutate(barrier.id = 1:length(km)) |>
  data.frame() |>
  subset(barrier.id \%in\%  # make predictions for barrier segments in cT
           seq(min(example_cT$barrier.id), max(example_cT$barrier.id, by = 1)))

predictions$kappa.hat <- predict(hwy1_w_model, se.fit = FALSE, 
                                 new.data = predictions)[,1]

# static map (no basemap) ----------------------------------------------
map <- mapPredictions(predictions, barrier.sf = hwy1_west)

## add animal tracks
example_sf <- example_id |>
  st_as_sf(coords=c("X","Y"), crs = st_crs(hwy1_west))
map <- mapPredictions(predictions, barrier.sf = hwy1_west,
                      tracks.sf = example_sf)

# static map (with basemap) --------------------------------------------
## default - landscape from ESRI and basemaps package
map <- mapPredictions(predictions, barrier.sf = hwy1_west, 
                      add.basemap = TRUE)
map

# interactive map ------------------------------------------------------
mapPredictions(predictions, barrier.sf = hwy1_west,
               interactive = TRUE)

}
