% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrossings.R
\name{plotCrossings}
\alias{plotCrossings}
\title{Plot Crossings}
\usage{
plotCrossings(
  permdata,
  original.data = NULL,
  barrier.sf = NULL,
  interactive.map = FALSE,
  static.map = FALSE,
  individual.plots = FALSE,
  add.basemap = FALSE
)
}
\arguments{
\item{permdata}{output from \link{prepPermeability} with \code{track} and
\code{barrier} objects, as stepwise dataframes.}

\item{original.data}{needed if making an interactive or static map with the
\code{mapview} or \code{ggplot2}/\code{ggspatial} packages. Should be a dataframe of
original data, containing columns for:
\code{ID}, \code{Time} (POSIXct format, date and time), and spatial coordinates
(\code{X}/\code{Y} - units in meters) - coordinate system should match barrier \code{sf}
object. Default NULL.}

\item{barrier.sf}{needed if making an interactive or static map with the
\code{mapview} or \code{ggplot2}/\code{ggspatial} packages. Should be an sf LINESTRING
object and coordinate reference system should match \code{original.data}. Default
NULL.}

\item{interactive.map}{if TRUE, will return interactive map(s) using the
\code{mapview} package. Default NULL.}

\item{static.map}{if TRUE, will return static map(s) using the \code{ggplot2} and
optionally the \code{ggspatial} package (if \code{add.basemap = TRUE}). Default FALSE.}

\item{individual.plots}{if TRUE, will return individual plots, for those
individuals that crossed the barrier - default FALSE (plots all tracks).}

\item{add.basemap}{if TRUE (default FALSE), will add a basemap for a
static map, using the \code{ggspatial} package. If your data has a large
spatial extent, it might take a couple minutes to compile the basemap.}
}
\value{
Plot(s) - if multiple individuals, will be returned as a list of
plots.
}
\description{
Function to plot crossings as output from either
\link{prepPermeability} or \link{simulatePermeability}. Defaults to basic
base R plots but for "real" data (e.g., data not simulated with
\code{simulatePermeability}) users can load additional packages for interactive
spatial maps (\code{mapview} package) or static spatial maps (\code{ggplot2} package),
optionally with a basemap (\code{ggspatial} package).
}
\examples{
require(permeability)
require(ggplot2)
require(ggspatial)
require(mapview)

data("example_id")
data("example_permdata")
data("hwy1_west")

# base R version -----------------------------

## simulated data
sim_tracks <- simulatePermeability(xmax = 30, ymax = 30, 
                                   N.steps = 50, theta.rho = 0.5, 
                                   step.shape = 5, step.scale = 10, 
                                   kappa = 0.2,
                                   n.tracks = 10,
                                   n.segments = 20,
                                   plot.track = FALSE) 

plotCrossings(permdata = sim_tracks)

## "real" data
plotCrossings(permdata = example_permdata)

# interactive map (mapview package) ------------------------------------
crossing_map <- plotCrossings(permdata = example_permdata, 
                              original.data = example_id,
                              barrier.sf = hwy1_west,
                              interactive.map = TRUE)

crossing_map

# static map (ggspatial and ggplot packages) --------------------------------

## just ggplot
crossing_map <- plotCrossings(permdata = example_permdata, 
                              original.data = example_id,
                              barrier.sf = hwy1_west,
                              static.map = TRUE)

crossing_map

## with basemap
crossing_map <- plotCrossings(permdata = example_permdata, 
                              original.data = example_id,
                              barrier.sf = hwy1_west,
                              static.map = TRUE,
                              add.basemap = TRUE)

crossing_map


}
