% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictPermeability.R
\name{predict.permfit}
\alias{predict.permfit}
\title{Prediction method for permeability fit}
\usage{
\method{predict}{permfit}(
  fit,
  new.data = NULL,
  se.fit = TRUE,
  conf.level = 0.95,
  bootstrap = FALSE,
  bootstrap.n = 1000,
  parallel = FALSE
)
}
\arguments{
\item{fit}{permeability fit object, output of \link{`fitPermeability`}}

\item{new.data}{an optional data frame if predictions are needed over new
values.}

\item{conf.level}{confidence level (default 95\%) .}

\item{bootstrap}{whether to get standard errors on estimates using
bootstrapping. Recommended if there are issues getting standard errors with
the Hessian with \link{`fitPermeability`}.}

\item{bootstrap.n}{number of iterations to run the bootstrapping for, if
selected. Defaults to 1000.}

\item{parallel}{if bootstrap = TRUE, whether to run in parallel using the
future.apply package or not. Recommended (bootstrapping can take 10+ min).}
}
\value{
data frame with columns \code{kappa.hat} - point estimate of
permeability,
\code{ci.low} and \code{ci.high} - confidence intervals around the kappa estimate
}
\description{
Prediction method for permeability fit
}
