% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{traffic_data}
\alias{traffic_data}
\title{Traffic data for HWY 1 West}
\format{
Data frame with columns for Date (POSIX UTC),
and traffic volume (number of vehicles),
}
\source{
GNWT
}
\usage{
data(traffic_data)
}
\description{
Daily traffic volume data for Highway 1 West, from a GNWT traffic counter.
}
\examples{
library(sf);library(mapview);library(dplyr)

# load caribou data
data(boreal_caribou)

# load highway data
data(hwy1_west)

# convert to sf structure
boreal_caribou_sf <- st_as_sf(boreal_caribou, 
                               coords = c("X","Y"), crs = 32611)

# map track highway data
boreal_tracks <- boreal_caribou_sf |> 
  group_by(ID) |>
  summarize(do_union=FALSE) |> 
  st_cast("LINESTRING")

mapview(boreal_tracks, zcol = "ID") + 
  mapview(hwy1_west, color = "red")

}
\keyword{data}
